+++ apps/web/src/components/Announcement/constants.ts
@@ -3,20 +3,24 @@ export const ANNOUNCEMENT_CONTENT = [
     key: 'MultiGalleryAnnouncement',
     title: 'Multi Gallery',
     description: 'You can now create multiple galleries and add them to your profile.',
+    date: '2023-03-14T04:56:26.154845Z',
   },
   {
     key: 'EmailUpsell',
     title: 'Email Upsell',
     description: 'You can now add your email address to your profile. ',
+    date: '2023-03-14T04:56:26.154845Z',
   },
   {
     key: 'MaintenanceFeb2023',
     title: 'Maintenance Feb 2023',
     description: 'We will be performing maintenance on the site on Feb 2023.',
+    date: '2023-03-14T04:56:26.154845Z',
   },
   {
     key: 'TwitterConnectionOnboardingUpsell',
     title: 'Connect Twitter',
     description: 'Connect your Twitter account to your profile.',
+    date: '2023-03-14T04:56:26.154845Z',
   },
 ];
